+++ application/src/main/resources/mappers/review/ReviewMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="core.application.reviews.repositories.mapper.ReviewMapper">
+<mapper namespace="core.application.reviews.repositories.mybatis.mappers.ReviewMapper">
 
     <insert id="saveNewReview" parameterType="map" useGeneratedKeys="true" keyProperty="review.reviewId" keyColumn="review_id">
         INSERT INTO review_table
